Update readme, code review fixes, and add test.
We did not tests deduplication or the requeued functionality. I added a
test for that. I updated the README because the tests are no longer a
script but rather a module. Removed the run-tests script because it is
not used any longer.

Fix tests where there were missing returns and stopListenings.

This sets up many listeners so if we miss stopping one it will
drastically change the outcome of the test.

Adding deadlock count.